Move dat provider to data provider class
